requirements.txt
Django==4.2.25
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
python-decouple==3.8
PyMySQL==1.1.0
Pillow==11.0.0
django-filter==25.1
gunicorn==23.0.0

# File compression and image processing
opencv-python==4.12.0.88
imageio==2.37.0
imageio-ffmpeg==0.6.0
numpy==2.2.6

# AWS S3 storage
boto3==1.40.59
django-storages==1.14.6

# WhatsApp messaging
twilio==9.4.1